refactor: move changed file picker to `commands`
